refactor(auction): 컨벤션에 맞게 네임 수정
