+++ src/main/java/com/bugzero/rarego/boundedContext/auction/app/AuctionSettleExpiredUseCase.java
@@ -3,8 +3,6 @@ package com.bugzero.rarego.boundedContext.auction.app;
 import com.bugzero.rarego.boundedContext.auction.domain.*;
 import com.bugzero.rarego.boundedContext.auction.out.AuctionOrderRepository;
 import com.bugzero.rarego.boundedContext.auction.out.AuctionRepository;
-import com.bugzero.rarego.global.exception.CustomException;
-import com.bugzero.rarego.global.response.ErrorType;
 import com.bugzero.rarego.shared.auction.event.AuctionEndedEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,11 +11,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class SettleOneAuctionUseCase {
+public class AuctionSettleExpiredUseCase {
 
     private final AuctionSettlementSupport support;
     private final AuctionRepository auctionRepository;
@@ -25,35 +25,36 @@ public class SettleOneAuctionUseCase {
     private final ApplicationEventPublisher eventPublisher;
 
     @Transactional
-    public void execute(Long auctionId) {
-        Auction auction = auctionRepository.findById(auctionId)
-                .orElseThrow(() -> new CustomException(ErrorType.AUCTION_NOT_FOUND));
+    public AuctionAutoResponseDto execute() {
+        LocalDateTime now = LocalDateTime.now();
+        List<Auction> auctions = support.findExpiredAuctions(now);
 
-        // 상태 검증
-        if (auction.getStatus() == AuctionStatus.ENDED) {
-            log.warn("경매 {}는 이미 종료되었습니다.", auctionId);
-            return;
+        int success = 0;
+        int fail = 0;
+        List<AuctionAutoResponseDto.SettlementDetail> details = new ArrayList<>();
+
+        for (Auction auction : auctions) {
+            try {
+                if (support.hasBids(auction.getId())) {
+                    handleSuccess(auction);
+                    Bid winningBid = support.findWinningBid(auction.getId());
+                    details.add(AuctionAutoResponseDto.SettlementDetail.success(
+                            auction.getId(),
+                            winningBid.getBidderId()
+                    ));
+                    success++;
+                } else {
+                    handleFail(auction);
+                    details.add(AuctionAutoResponseDto.SettlementDetail.failed(auction.getId()));
+                    fail++;
+                }
+            } catch (Exception e) {
+                log.error("경매 낙찰 처리 실패 - auctionId: {}", auction.getId(), e);
+                fail++;
+            }
         }
 
-        if (auction.getStatus() != AuctionStatus.IN_PROGRESS) {
-            log.warn("경매 {}는 진행 중 상태가 아닙니다. 현재 상태: {}", auctionId, auction.getStatus());
-            throw new CustomException(ErrorType.AUCTION_NOT_IN_PROGRESS);
-        }
-
-        // 시간 검증
-        if (auction.getEndTime().isAfter(LocalDateTime.now())) {
-            log.warn("경매 {}는 아직 종료 시간이 아닙니다. 종료 시간: {}", auctionId, auction.getEndTime());
-            throw new CustomException(ErrorType.AUCTION_NOT_IN_PROGRESS);
-        }
-
-        // 입찰 여부에 따라 처리
-        if (support.hasBids(auctionId)) {
-            handleSuccess(auction);
-            log.info("경매 {} 낙찰 처리 완료", auctionId);
-        } else {
-            handleFail(auction);
-            log.info("경매 {} 유찰 처리 완료", auctionId);
-        }
+        return AuctionAutoResponseDto.from(now, auctions, success, fail, details);
     }
 
     private void handleFail(Auction auction) {
